# connection/libvirt_qemu: Mitigate a CPU hammering active wait loop