##### SUMMARY
<!--- Describe the change below, including rationale and design decisions -->

Once the `libvirt_qemu` connection plugin has spawned a command via the qemu guest agent, it needs to wait for it to finish.
This is done in an active wait loop.

However, without any sleep, this wait loop consumes a lot of CPU.

On my machine, the following command:
```
ansible archlinux-vm -m command -a "sleep 10"
```
makes the `qemu-ga` of my VM consume 20% of CPU and `systemd-journald` consume 7% of CPU because all qemu guest agent commands are logged.

Adding a fixed value sleep time in each loop iteration could make tasks execution run slower, especially if there are a lot of quick ones.

In order to not constantly consume a CPU for long-running tasks without increasing the latency on quick ones, this change introduces an exponentially increasing sleep in the active wait loop.
It guarantees that the latency increase it bounded up to 5% of the task execution time while drastically reducing the CPU consumption.

Here are the impacts of this patch on the following command:
```
ansible archlinux-vm -m command -a "sleep 10"
```

| Measure                       |  Before  |  After   |
| :---------------------------- | -------: | -------: |
| Elapsed time                  | 11220 ms | 11389 ms |
| CPU time ansible host         |  3539 ms |  2655 ms |
| CPU libvirt VM ansible target |  3826 ms |   271 ms |

We can see that the CPU consumption has been drastically reduced without increasing the elapsed time too much.

##### ISSUE TYPE
<!--- Pick one below and delete the rest -->
- Bugfix Pull Request

##### COMPONENT NAME
<!--- Write the short name of the module, plugin, task or feature below -->
`connection/libvirt_qemu`

##### ADDITIONAL INFORMATION
<!--- Include additional information to help people understand the change here -->
<!--- A step-by-step reproduction of the problem is helpful if there is no related issue -->

<!--- Paste verbatim command output below, e.g. before and after your change -->
Here is the benchmark program I used to produce the above figures:
```bash
#!/usr/bin/bash
set -euo pipefail

# Works only with cgroup v2
terminal_cgroup=$(awk -F: '$1 == "0" {print $3}' < "/proc/self/cgroup")
# Works only if there’s a single VM running
vm_cgroup=$(awk -F: '$1 == "0" {print $3}' < "/proc/$(pgrep qemu)/cgroup")

before_date_nsec=$(date +%s%N)
before_terminal_usage_usec=$(awk '$1 == "usage_usec" {print $2}' < "/sys/fs/cgroup${terminal_cgroup}/cpu.stat")
before_vm_usage_usec=$(      awk '$1 == "usage_usec" {print $2}' < "/sys/fs/cgroup${vm_cgroup}/cpu.stat")

ansible archlinux-vm -m command -a "sleep 10"

after_date_nsec=$(date +%s%N)
after_terminal_usage_usec=$(awk '$1 == "usage_usec" {print $2}' < "/sys/fs/cgroup${terminal_cgroup}/cpu.stat")
after_vm_usage_usec=$(      awk '$1 == "usage_usec" {print $2}' < "/sys/fs/cgroup${vm_cgroup}/cpu.stat")

echo "Elapsed time: $(((after_date_nsec-before_date_nsec)/1000000))ms"
echo "CPU time ansible host: $(((after_terminal_usage_usec-before_terminal_usage_usec)/1000))ms"
echo "CPU libvirt VM ansible target: $(((after_vm_usage_usec-before_vm_usage_usec)/1000))ms"
```